refactor: MainFooter 컴포넌트의 계층 이동 (organisms)
