+++ src/components/index.ts
@@ -10,11 +10,11 @@ export { Item } from 'components/atoms/Item';
 export { Accordion } from 'components/molecules/Accordion';
 export { BrandLists } from 'components/molecules/BrandLists';
 export { ItemLists } from 'components/molecules/ItemLists';
-export { MainFooter } from 'components/molecules/MainFooter';
 
 // Organisms
 export { CustomerServiceInfo } from 'components/organisms/CustomerServiceInfo';
 export { Faq } from 'components/organisms/Faq';
+export { MainFooter } from 'components/organisms/MainFooter';
 export { MenuSelector } from 'components/organisms/MenuSelector';
 export { SlideMenu } from 'components/organisms/SlideMenu';
 export { NavBar } from 'components/organisms/NavBar';
